dashboard/requirements.txt
streamlit==1.15.0
psycopg2-binary==2.9.3
pandas==1.4.3
